pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling>=1.28.0"]

[project]
name = "license-manager"
version = "1.0.01"

authors = [{ name = "HRBN" }]
description = "Backend for the license manager app"
license = { text = "Proprietary" }
readme = "README.md"
requires-python = ">=3.13.11"

dependencies = [
    # API & server
    "fastapi==0.135.2",
    "uvicorn==0.42.0",
    "gunicorn==25.3.0",

    # Config & validation
    "pydantic==2.12.5",
    "pydantic[email]==2.12.5",
    "pydantic-settings==2.13.1",

    # Async tasks & messaging
    "celery==5.6.3",
    "pika==1.3.2",   # optional direct RabbitMQ usage alongside Celery/kombu

    # Auth / security
    "fastapi_keycloak==1.1.1",

    # Pagination
    "fastapi-pagination==0.15.11",

    # Azure Key Vault (secrets/config)
    "azure-identity==1.25.3",
    "azure-keyvault-secrets==4.10.0",

    # Email (for alerts/notifications)
    "aiosmtplib==5.1.0",
    "mjml-python==1.3.7",

    # Database: SQLAlchemy async + psycopg (PostgreSQL)
    "sqlalchemy==2.0.48",
    "psycopg[binary]==3.3.3",
    "greenlet==3.3.2",
    "alembic==1.18.4",
    "sqlmodel==0.0.37",

    # Logging
    "python-json-logger==4.1.0",
] # --- runtime deps ---

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "mypy==1.19.1",
    "ruff==0.15.2",
    "types-requests==2.33.0.20260327",
    "pytest-cov==7.1.0",
    "black==26.3.1",
]

[project.urls]
Homepage = "https://hrbn.pl"

[tool.hatch.build.targets.wheel]
packages = ["app/license_manager"]

# ---------- Ruff ----------
[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
ignore = []
select = ["E", "F", "B", "UP", "I"]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = [
    "fastapi.Query",
    "fastapi.Depends",
    "fastapi.Body",
    "fastapi.Path",
]

# ---------- mypy ----------
[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
plugins = ["pydantic.mypy"]
python_version = "3.13.11"
strict = true
warn_redundant_casts = true
warn_unused_ignores = true

# ---------- pytest ----------
[tool.pytest.ini_options]
addopts = "-q --cov=app/license_manager"
asyncio_mode = "auto"
testpaths = ["tests"]
